Renamed  files accordingly
- extracted unrelated methods to separate files
