+++ lib/Nonce.php
